Refactor dto (#12)
* Refactor dtos to new SDK package

* Add authConfig support to CogniteClient

* Start read events and aggregation support

* Add aggregate support to event. With test.

* Events list migrated to new sdk

* Set default max worker threads to 24. Do not create new ForkJoinPool unless necessary.

* Add generic read by id method

* Refactor resource types to the SDK

* Add parsing methods

* Add asset generator

* Add some default client config settings.

* Add checks for externalId and circular references

* Add check duplicates asset method

* Add method skaffold

* Add method skaffold for upsertViaDelete

* Add item function for delete operations.

* Restore the delete items writer. Need it for propagating it to the delete worker.

* Add preconditions skaffold.

* Add topological sorting

* Finish assets upsert

* Reduce default max workers to 16.

* Implement upsertViaDelete

* Implement upsert via update and create

* Add relationships api

* Add relationships api tests

* Set default max worker threads to 12

* Add upsert / replace test

* Fix. Parse targetExternalId when reading relationships

* Fix tests.

* Add timeseries headers to the SDK

* Add sequences header/metadata support

* Add read data set by id service.

* Add read data set by id to api base.

* Add aggregate data set support.

* Add skaffold for time series data and sequences data.

* Start setting up delete sequences rows

* Add delete items parameters (configuration) to UpsertItems

* Add labels

* Add labels

* Add executor and client handling to read iterators

* Start sequence rows

* Finish upsert implementation

* Finish sequences rows support.

* Add logging

* Add counter functions

* New split controller for upserting sequences rows. Trigger on max cell count and max character count.

* Start data points upsert implementation.

* Continue data points upsert methods

* Continue data points upsert

* group by id

* Fix. Do not add an empty row to a batch

* Add split and upsert

* Completed upsert

* Add delete data points support

* Start retrieve function.

* Fix. Wrong sequence of iterators for retrieve

* Clean up imports

* Finish retrieve and split methods

* Add test for data points.

* Set max default workers to 8

* Optimize logging.

* Set default max workers to 8

* Cleanup

* Add ts point json parser

* Add read latest data point service

* Adjust data points test load

* Start data points latest

* Add retrieve latest and first. Optimize retrieve data points.

* Add logging of api latency

* Refactor write files service (method name)

* Set max workers for file uploads to 20

* Add support for the files resource type

* Refactor file service method

* Start upsert file header

* Start upsert header service

* Add split and batch methods

* Complete file upload. Start file download

* Complete file upload. Start file download

* Start file test

* Complete the file binary download

* Add file container merge method

* Complete files support and tests

* Start implementation of Raw tables

* Clean up imports

* Add Raw table api. Improve logger registration

* Add Raw db api.

* Clean up comments

* Add support for security categories

* Fix delete to a long array

* start raw rows support

* Start list rows

* Add testing for raw db and tables. Fix test for security categories

* Add list with custom partitions

* Add read raw cursors.

* Add list rows

* Add list with auto cursors.

* Start skeleton for upsert and delete

* Complete upsert and delete methods. Add rows to test.

* Fix. list raw rows.

* Add delete rows to test

* Add retrieve row

* Add new assets write and list to Beam.

* Add ts header list and upsert new sdk

* Add list and upsert file headers using the new sdk

* Add retrieve file binary via sdk

* Add list, retrieve, upsert for sequences. Add retrieve for assets, events, timeseries.

* Add list, retrieve, upsert for sequences. Add retrieve for assets, events, timeseries.

* Add list and upsert for data sets, labels and relationships via sdk

* Add retrieve for data sets and relationships via sdk

* Add list and upsert sequences body

* Add list ts datapoints proto via sdk

* Add streaming / item timestamp support.

* Upsert datapoints via sdk. Increase max batch size to 200k.

* Read file by id via sdk. Remove old read by id fn.

* Refactor delete items to use the new sdk

* Remove old items iterator

* Remove old partitions fn

* Delete raw rows to use the new sdk

* Refactor delete raw rows and sequence rows to use the new sdk

* Refactor read raw databases to use the new sdk

* Refactor read raw tables to use the new sdk

* Refactor read raw rows to use the new sdk

* Refactor file upload to use new sdk. Remove unused files.

* Improve logging

* Add raw rows method with columns list support

* Add contextualization skaffold

* Fix. Parse the name attribute for raw db and tables

* Add support for columns and request parameters

* Guard against null pointer on the size parameter

* Start entity matching implementation

* Refactor entity matching results object.

* Update release notes.

* Fix. Do not use "ignoreUnknownIds" for file binaries.

* Add first method

* Completed implementing the predict methods

* Add javadoc

* Add delete model

* Move entity matching requests from playground to v1

* Finalize entity create and test.

* Refactor match entities to use the new sdk methods.

* Start sdk for interactive pnid

* Fix. Null handling of GCS blob when checking content lenght.

* Fix. Null handling of GCS blob when uploading the temp binary.

* Fix. Refactor test to use annotations.

* Add logging of content length.

* Start refactor file binary download

* Finished refactor file download

* Start fix for TS frequency

* Update logging with batch identifiers.

* Fix. Read datapoints. Don't check aggregates for string time series.

* Fix. Potential division by zero

* Fix. floating point format

* Fix. wrong logging calculation

* Refactor interactive P&ID to use new SDK.

* Fix. Grayscale parameter. Add sdk test for interactive pnid.

+++ src/main/java/com/cognite/client/servicesV1/executor/FileBinaryRequestExecutor.java
@@ -16,7 +16,7 @@
 
 package com.cognite.client.servicesV1.executor;
 
-import com.cognite.beam.io.dto.FileBinary;
+import com.cognite.client.dto.FileBinary;
 import com.cognite.client.servicesV1.ConnectorConstants;
 import com.cognite.client.servicesV1.ResponseBinary;
 import com.google.auto.value.AutoValue;
@@ -82,9 +82,8 @@ public abstract class FileBinaryRequestExecutor {
             504     // gateway timeout
     );
 
-    private static final int DEFAULT_CPU_MULTIPLIER = 10;
-    private static final ForkJoinPool DEFAULT_POOL = new ForkJoinPool(Runtime.getRuntime().availableProcessors()
-            * DEFAULT_CPU_MULTIPLIER);
+    private static final int DEFAULT_NUM_WORKERS = 10;
+    private static final ForkJoinPool DEFAULT_POOL = new ForkJoinPool(DEFAULT_NUM_WORKERS);
 
     protected final static Logger LOG = LoggerFactory.getLogger(FileBinaryRequestExecutor.class);
 
@@ -303,7 +302,7 @@ public abstract class FileBinaryRequestExecutor {
                             String.format("%.2f", response.body().contentLength() / (1024d * 1024d)));
                     return FileBinary.newBuilder()
                             .setBinary(ByteString.readFrom(response.body().byteStream()))
-                            .setContentLength(Int64Value.of(response.body().contentLength()))
+                            .setContentLength(response.body().contentLength())
                             .build();
                 }
             } catch (Exception e) {
@@ -545,7 +544,7 @@ public abstract class FileBinaryRequestExecutor {
 
             return FileBinary.newBuilder()
                     .setBinaryUri(fileURI.toString())
-                    .setContentLength(Int64Value.of(response.body().contentLength()))
+                    .setContentLength(response.body().contentLength())
                     .build();
         } else if (null != getTempStoragePath().getScheme()
                 && getTempStoragePath().getScheme().equalsIgnoreCase("file")) {
@@ -564,7 +563,7 @@ public abstract class FileBinaryRequestExecutor {
 
             return FileBinary.newBuilder()
                     .setBinaryUri(tempFilePath.toUri().toString())
-                    .setContentLength(Int64Value.of(response.body().contentLength()))
+                    .setContentLength(response.body().contentLength())
                     .build();
         } else {
             throw new IOException("Temp storage location not supported: " + getTempStoragePath());
@@ -655,6 +654,12 @@ public abstract class FileBinaryRequestExecutor {
         public void writeTo(@NotNull BufferedSink bufferedSink) throws IOException {
             if (null != fileURI.getScheme() && fileURI.getScheme().equalsIgnoreCase("gs")) {
                 Blob blob = getBlob(fileURI);
+                if (null == blob) {
+                    LOG.error("Looks like the GCS blob is null/does not exist. File URI: {}",
+                            fileURI.toString());
+                    throw new IOException(String.format("Looks like the GCS blob is null/does not exist. File URI: %s",
+                            fileURI.toString()));
+                }
                 blob.downloadTo(bufferedSink.outputStream());
                 if (deleteTempFile) {
                     blob.delete();
@@ -673,7 +678,7 @@ public abstract class FileBinaryRequestExecutor {
                         Files.delete(Paths.get(fileURI));
                     }
                 } catch (Exception e) {
-                    new IOException(e);
+                    throw new IOException(e);
                 }
             } else {
                 throw new IOException("URI is unsupported: " + fileURI.toString());
@@ -685,7 +690,12 @@ public abstract class FileBinaryRequestExecutor {
             if (fileURI.getScheme().equalsIgnoreCase("gs")) {
                 try {
                     Blob blob = getBlob(fileURI);
-                    contentLength = blob.getSize();
+                    if (null == blob) {
+                        LOG.warn("Looks like the GCS blob is null/does not exist. File URI: {}",
+                                fileURI.toString());
+                    } else {
+                        contentLength = blob.getSize();
+                    }
                 } catch (IOException e) {
                     LOG.warn(e.getMessage());
                 }
